requirements.txt
pandas
numpy
pyspark
boto3
psycopg2-binary